feat: add unit tests of risk stewards

+++ contracts/interfaces/IRiskSteward.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.25;
 interface IRiskSteward {
     function decodeAdditionalData(
         bytes calldata additionalData
-    ) external pure returns (address underlying, uint16 destChainId);
+    ) external pure returns (address underlying, uint32 destChainId);
 
     function processUpdate(uint256 updateId, bytes memory newValue, string memory updateType, address market) external;
